requirements.txt
streamlit
sqlalchemy
psycopg2-binary
pandas
Pillow